Module Default.Latest

This module contains updated versions of node constructors that were kept stable when the node changed. For every function in this module, there's an equally-named function outside this module. The function outside this module will stay stable, whereas the function inside this module will adapt potential upcoming new compiler features. Only use a function in this module, if the equally-named one outside this module is missing a feature you need.

val ppat_construct : loc:Astlib.Location.t -> Astlib.Longident.t Astlib.Location.loc -> (string Astlib.Location.loc list * Astlib.Ast_500.Parsetree.pattern) option -> Astlib.Ast_500.Parsetree.pattern

val constructor_declaration : loc:Astlib.Location.t -> name:string Astlib.Location.loc -> vars:string Astlib.Location.loc list -> args:Astlib.Ast_500.Parsetree.constructor_arguments -> res:Astlib.Ast_500.Parsetree.core_type option -> unit -> Astlib.Ast_500.Parsetree.constructor_declaration