Module Ppxlib.Context_free

Context free rewriting, to define local rewriting rules that will all be applied at once by the driver.

module Rule : sig ... end

Local rewriting rules.

class map_top_down : ?expect_mismatch_handler:{Expect_mismatch_handler}2.t -> ?generated_code_hook: {Generated_code_hook}1.t -> ?embed_errors:bool -> Rule.t list -> object ... end