Module Generator.V2

val make : ?attributes:Attribute.packed list -> ?deps:t list -> ('f, 'output_ast) Args.t -> (ctxt:Expansion_context.Deriver.t -> 'input_ast -> 'f) -> ('output_ast, 'input_ast) t

Same as Generator.make, but the generator has access to an expansion context.

val make_noarg : ?attributes:Attribute.packed list -> ?deps:t list -> (ctxt:Expansion_context.Deriver.t -> 'input_ast -> 'output_ast) -> ('output_ast, 'input_ast) t

Same as Generator.make_noarg, but the generator has access to an expansion context.