Module Expansion_context.Base

type t

Type for the location independent parts of the expansion context

val top_level : omp_config:Migrate_parsetree.Driver.config -> file_path:string -> t

Build a new base context at the top level of the given file with the given ocaml-mirgate-parsetree configuration.

val enter_expr : t -> t

Proxy functions to update the wrapped code path. See code_path.mli for details.

val enter_module : loc:Location.t -> string -> t -> t

val enter_value : loc:Location.t -> string -> t -> t