Module Expansion_context.Deriver

type t

Type of expansion contexts for derivers

val derived_item_loc : t -> Location.t

Return the location of the item to which the deriver is being applied

val code_path : t -> Code_path.t

Return the code path for the given context

val omp_config : t -> Migrate_parsetree.Driver.config

Return the ocaml-migrate-parsetree configuration for the given expansion context

val with_loc_and_path : (loc:Location.t -> path:string -> 'a) -> ctxt:t -> 'a

Wrap a fun ~loc ~path into a fun ~ctxt

val inline : t -> bool

Whether the derived code is going to be inlined in the source