Module Location.Error

type t

val make : loc:t -> Base.String.t -> sub:(t * Base.String.t) Base.List.t -> t

val createf : loc:t -> ('a, Base.Unit.t, Base.String.t, t) Stdlib.format4 -> 'a

val message : t -> Base.String.t

val set_message : t -> Base.String.t -> t

val register_error_of_exn : (Base.Exn.t -> t Base.Option.t) -> Base.Unit.t

Register an exception handler. Exception registered this way will be properly displayed by report_exception.

val of_exn : Base.Exn.t -> t Base.Option.t

val to_extension : t -> string Ocaml_common.Location.loc * Migrate_parsetree.Ast_408.Parsetree.payload

Convert an error to an extension point. The compiler recognizes this and displays the error properly.