Module Pp_ast.Config

type t

Type for AST pretty-printing config

val make : ?show_attrs:bool -> ?show_locs:bool -> ?loc_mode:[ `Short | `Full ] -> unit -> t

Create a custom pretty-printing config. Default values are the ones that are used when no configuration is passed to the pretty-printers defined in Pp_ast.

parameter ?show_attrs
controls whether attributes are shown or hidden. Defaults to false.

parameter ?show_loc
controls whether locations are shown or hidden. Defaults to false.

parameter ?loc_mode
controls how locations are shown if they are shown at all. Defaults to `Short.

When set to `Short, locations are displayed as "l1c6..l2c2" for multiline locations and as "l1c6..12" for single line locations. Ghost locations are suffixed with a "(g)".

When set to `Full, locations are displayed as any other record would be.