Module Ppxlib_ast__Versions

Abstracting an OCaml frontend

module type Ast = sig ... end

Abstract view of a version of an OCaml Ast

type 'a _types = 'a constraint 'a = < structure : _; signature : _; toplevel_phrase : _; core_type : _; expression : _; pattern : _; case : _; type_declaration : _; type_extension : _; extension_constructor : _; >

type _ witnesses = private ..

A version of the OCaml frontend packs the ast with type witnesses so that equalities can be recovered dynamically.

type _ migration_info

migration_info is an opaque type that is used to generate migration functions.

module type OCaml_version = sig ... end

An OCaml frontend versions an Ast, version number and some witnesses for conversion.

Concrete frontend instances

module OCaml_402 : OCaml_version with module Ast = Migrate_parsetree.Ast_402

module OCaml_403 : OCaml_version with module Ast = Migrate_parsetree.Ast_403

module OCaml_404 : OCaml_version with module Ast = Migrate_parsetree.Ast_404

module OCaml_405 : OCaml_version with module Ast = Migrate_parsetree.Ast_405

module OCaml_406 : OCaml_version with module Ast = Migrate_parsetree.Ast_406

module OCaml_407 : OCaml_version with module Ast = Migrate_parsetree.Ast_407

module OCaml_408 : OCaml_version with module Ast = Migrate_parsetree.Ast_408

module OCaml_409 : OCaml_version with module Ast = Migrate_parsetree.Ast_409

module OCaml_410 : OCaml_version with module Ast = Migrate_parsetree.Ast_410

module OCaml_411 : OCaml_version with module Ast = Migrate_parsetree.Ast_411

module OCaml_current = OCaml_411

Convenience definitions

module Convert : functor (A : OCaml_version) -> functor (B : OCaml_version) -> sig ... end

Module level migration