Ppxlib's user manual

Overview

This is the user manual and api for ppxlib, the core of the ppx meta-programming system for OCaml and its derivatives such as Reason. This manual is aimed at both users and authors of ppx rewriters and contains everything one should know in order to use or write ppx rewriters.

It is assumed in this manual that the user is familiar with the Dune build system. In particular, all the examples in this manual referring to the build system will present Dune configurations files and commands. It is possible to use ppxlib with other build systems, however this is not covered by this manual.

Manual

The manual consists of three main sections:

What is ppx

Ppxlib for end users

Ppxlib for plugin authors

API

The API exposes the following modules:

Ppxlib Standard library for ppx rewriters

Ppxlib_ast

Astlib

Ppxlib_metaquot

Ppxlib_metaquot_lifters

Ppxlib_print_diff

Ppxlib_runner

Ppxlib_runner_as_ppx

Stdppx

Ppxlib_traverse

Ppxlib_traverse_builtins