Module Location.Error

type t

The location error type. It contains a located main message and a (possibly empty) list of located submessages.

val is_well_formed : t -> bool

A location error constructed via make is always well-formed. A malformed location error is a value of type location_report on compilers >= 4.08, whose kind is different from Report_error. Notice that location_report does not explicitly form part of Astlib.

val main_msg : t -> string loc

Get the located error main message.

val sub_msgs : t -> string loc list

Get the located error sub-messages.

val set_main_msg : t -> string -> t

Set the text of the error's main message. The location stays as is.

val set_main_loc : t -> t -> t

Set the location of the error's main message. The text satys as is.

val make : sub:string loc list -> string loc -> t

Construct a location error.

val of_exn : exn -> t option

Turn an exception into a location error, if possible.