Module Ppxlib

Standard library for ppx rewriters

include module type of struct include Ocaml_shadow end with module Ast_helper := Ocaml_shadow.Ast_helper with module Asttypes := Ocaml_shadow.Asttypes with module Docstrings := Ocaml_shadow.Docstrings with module Identifiable := Ocaml_shadow.Identifiable with module Lexer := Ocaml_shadow.Lexer with module Location := Ocaml_shadow.Location with module Longident := Ocaml_shadow.Longident with module Parse := Ocaml_shadow.Parse with module Parser := Ocaml_shadow.Parser with module Parsetree := Ocaml_shadow.Parsetree with module Pprintast := Ocaml_shadow.Pprintast with module Syntaxerr := Ocaml_shadow.Syntaxerr

module Do_not_use_directly = Ocaml_shadow.Do_not_use_directly

module Afl_instrument = Ocaml_shadow.Afl_instrument

module Alias_analysis = Ocaml_shadow.Alias_analysis

module Allocated_const = Ocaml_shadow.Allocated_const

module Annot = Ocaml_shadow.Annot

module Arch = Ocaml_shadow.Arch

module Arg_helper = Ocaml_shadow.Arg_helper

module Asmgen = Ocaml_shadow.Asmgen

module Asmlibrarian = Ocaml_shadow.Asmlibrarian

module Asmlink = Ocaml_shadow.Asmlink

module Asmpackager = Ocaml_shadow.Asmpackager

module Ast_invariants = Ocaml_shadow.Ast_invariants

module Ast_iterator = Ocaml_shadow.Ast_iterator

module Ast_mapper = Ocaml_shadow.Ast_mapper

module Attr_helper = Ocaml_shadow.Attr_helper

module Augment_specialised_args = Ocaml_shadow.Augment_specialised_args

module Available_regs = Ocaml_shadow.Available_regs

module Backend_intf = Ocaml_shadow.Backend_intf

module Backend_var = Ocaml_shadow.Backend_var

module Binutils = Ocaml_shadow.Binutils

module Branch_relaxation = Ocaml_shadow.Branch_relaxation

module Branch_relaxation_intf = Ocaml_shadow.Branch_relaxation_intf

module Btype = Ocaml_shadow.Btype

module Build_export_info = Ocaml_shadow.Build_export_info

module Build_path_prefix_map = Ocaml_shadow.Build_path_prefix_map

module Builtin_attributes = Ocaml_shadow.Builtin_attributes

module Bytegen = Ocaml_shadow.Bytegen

module Bytelibrarian = Ocaml_shadow.Bytelibrarian

module Bytelink = Ocaml_shadow.Bytelink

module Bytepackager = Ocaml_shadow.Bytepackager

module Bytesections = Ocaml_shadow.Bytesections

module CSE = Ocaml_shadow.CSE

module CSEgen = Ocaml_shadow.CSEgen

module CamlinternalMenhirLib = Ocaml_shadow.CamlinternalMenhirLib

module Ccomp = Ocaml_shadow.Ccomp

module Clambda = Ocaml_shadow.Clambda

module Clambda_primitives = Ocaml_shadow.Clambda_primitives

module Clflags = Ocaml_shadow.Clflags

module Closure = Ocaml_shadow.Closure

module Closure_conversion = Ocaml_shadow.Closure_conversion

module Closure_conversion_aux = Ocaml_shadow.Closure_conversion_aux

module Closure_element = Ocaml_shadow.Closure_element

module Closure_id = Ocaml_shadow.Closure_id

module Closure_middle_end = Ocaml_shadow.Closure_middle_end

module Closure_offsets = Ocaml_shadow.Closure_offsets

module Closure_origin = Ocaml_shadow.Closure_origin

module Cmi_format = Ocaml_shadow.Cmi_format

module Cmm = Ocaml_shadow.Cmm

module Cmm_helpers = Ocaml_shadow.Cmm_helpers

module Cmmgen = Ocaml_shadow.Cmmgen

module Cmmgen_state = Ocaml_shadow.Cmmgen_state

module Cmo_format = Ocaml_shadow.Cmo_format

module Cmt2annot = Ocaml_shadow.Cmt2annot

module Cmt_format = Ocaml_shadow.Cmt_format

module Cmx_format = Ocaml_shadow.Cmx_format

module Cmxs_format = Ocaml_shadow.Cmxs_format

module Coloring = Ocaml_shadow.Coloring

module Comballoc = Ocaml_shadow.Comballoc

module Compenv = Ocaml_shadow.Compenv

module Compilation_unit = Ocaml_shadow.Compilation_unit

module Compile = Ocaml_shadow.Compile

module Compile_common = Ocaml_shadow.Compile_common

module Compilenv = Ocaml_shadow.Compilenv

module Compmisc = Ocaml_shadow.Compmisc

module Compute_ranges = Ocaml_shadow.Compute_ranges

module Compute_ranges_intf = Ocaml_shadow.Compute_ranges_intf

module Config = Ocaml_shadow.Config

module Consistbl = Ocaml_shadow.Consistbl

module Convert_primitives = Ocaml_shadow.Convert_primitives

module Ctype = Ocaml_shadow.Ctype

module Datarepr = Ocaml_shadow.Datarepr

module Deadcode = Ocaml_shadow.Deadcode

module Debuginfo = Ocaml_shadow.Debuginfo

module Depend = Ocaml_shadow.Depend

module Dll = Ocaml_shadow.Dll

module Domainstate = Ocaml_shadow.Domainstate

module Effect_analysis = Ocaml_shadow.Effect_analysis

module Emit = Ocaml_shadow.Emit

module Emitaux = Ocaml_shadow.Emitaux

module Emitcode = Ocaml_shadow.Emitcode

module Env = Ocaml_shadow.Env

module Envaux = Ocaml_shadow.Envaux

module Errors = Ocaml_shadow.Errors

module Export_id = Ocaml_shadow.Export_id

module Export_info = Ocaml_shadow.Export_info

module Export_info_for_pack = Ocaml_shadow.Export_info_for_pack

module Expunge = Ocaml_shadow.Expunge

module Extract_projections = Ocaml_shadow.Extract_projections

module Find_recursive_functions = Ocaml_shadow.Find_recursive_functions

module Flambda = Ocaml_shadow.Flambda

module Flambda_invariants = Ocaml_shadow.Flambda_invariants

module Flambda_iterators = Ocaml_shadow.Flambda_iterators

module Flambda_middle_end = Ocaml_shadow.Flambda_middle_end

module Flambda_to_clambda = Ocaml_shadow.Flambda_to_clambda

module Flambda_utils = Ocaml_shadow.Flambda_utils

module Freshening = Ocaml_shadow.Freshening

module Genprintval = Ocaml_shadow.Genprintval

module Id_types = Ocaml_shadow.Id_types

module Ident = Ocaml_shadow.Ident

module Import_approx = Ocaml_shadow.Import_approx

module Includeclass = Ocaml_shadow.Includeclass

module Includecore = Ocaml_shadow.Includecore

module Includemod = Ocaml_shadow.Includemod

module Inconstant_idents = Ocaml_shadow.Inconstant_idents

module Initialize_symbol_to_let_symbol = Ocaml_shadow.Initialize_symbol_to_let_symbol

module Inline_and_simplify = Ocaml_shadow.Inline_and_simplify

module Inline_and_simplify_aux = Ocaml_shadow.Inline_and_simplify_aux

module Inlining_cost = Ocaml_shadow.Inlining_cost

module Inlining_decision = Ocaml_shadow.Inlining_decision

module Inlining_decision_intf = Ocaml_shadow.Inlining_decision_intf

module Inlining_stats = Ocaml_shadow.Inlining_stats

module Inlining_stats_types = Ocaml_shadow.Inlining_stats_types

module Inlining_transforms = Ocaml_shadow.Inlining_transforms

module Instruct = Ocaml_shadow.Instruct

module Int_replace_polymorphic_compare = Ocaml_shadow.Int_replace_polymorphic_compare

module Interf = Ocaml_shadow.Interf

module Internal_variable_names = Ocaml_shadow.Internal_variable_names

module Interval = Ocaml_shadow.Interval

module Invariant_params = Ocaml_shadow.Invariant_params

module Lambda = Ocaml_shadow.Lambda

module Lift_code = Ocaml_shadow.Lift_code

module Lift_constants = Ocaml_shadow.Lift_constants

module Lift_let_to_initialize_symbol = Ocaml_shadow.Lift_let_to_initialize_symbol

module Linear = Ocaml_shadow.Linear

module Linear_format = Ocaml_shadow.Linear_format

module Linearize = Ocaml_shadow.Linearize

module Linkage_name = Ocaml_shadow.Linkage_name

module Linscan = Ocaml_shadow.Linscan

module Liveness = Ocaml_shadow.Liveness

module Load_path = Ocaml_shadow.Load_path

module Local_store = Ocaml_shadow.Local_store

module Mach = Ocaml_shadow.Mach

module Main = Ocaml_shadow.Main

module Main_args = Ocaml_shadow.Main_args

module Maindriver = Ocaml_shadow.Maindriver

module Makedepend = Ocaml_shadow.Makedepend

module Matching = Ocaml_shadow.Matching

module Meta = Ocaml_shadow.Meta

module Misc = Ocaml_shadow.Misc

module Mtype = Ocaml_shadow.Mtype

module Mutable_variable = Ocaml_shadow.Mutable_variable

module Numbers = Ocaml_shadow.Numbers

module Opcodes = Ocaml_shadow.Opcodes

module Oprint = Ocaml_shadow.Oprint

module Optcompile = Ocaml_shadow.Optcompile

module Opterrors = Ocaml_shadow.Opterrors

module Optmain = Ocaml_shadow.Optmain

module Optmaindriver = Ocaml_shadow.Optmaindriver

module Outcometree = Ocaml_shadow.Outcometree

module Parameter = Ocaml_shadow.Parameter

module Parmatch = Ocaml_shadow.Parmatch

module Pass_wrapper = Ocaml_shadow.Pass_wrapper

module Path = Ocaml_shadow.Path

module Patterns = Ocaml_shadow.Patterns

module Persistent_env = Ocaml_shadow.Persistent_env

module Pparse = Ocaml_shadow.Pparse

module Predef = Ocaml_shadow.Predef

module Primitive = Ocaml_shadow.Primitive

module Printast = Ocaml_shadow.Printast

module Printclambda = Ocaml_shadow.Printclambda

module Printclambda_primitives = Ocaml_shadow.Printclambda_primitives

module Printcmm = Ocaml_shadow.Printcmm

module Printinstr = Ocaml_shadow.Printinstr

module Printlambda = Ocaml_shadow.Printlambda

module Printlinear = Ocaml_shadow.Printlinear

module Printmach = Ocaml_shadow.Printmach

module Printpat = Ocaml_shadow.Printpat

module Printtyp = Ocaml_shadow.Printtyp

module Printtyped = Ocaml_shadow.Printtyped

module Proc = Ocaml_shadow.Proc

module Profile = Ocaml_shadow.Profile

module Projection = Ocaml_shadow.Projection

module Rec_check = Ocaml_shadow.Rec_check

module Ref_to_variables = Ocaml_shadow.Ref_to_variables

module Reg = Ocaml_shadow.Reg

module Reg_availability_set = Ocaml_shadow.Reg_availability_set

module Reg_with_debug_info = Ocaml_shadow.Reg_with_debug_info

module Reload = Ocaml_shadow.Reload

module Reloadgen = Ocaml_shadow.Reloadgen

module Remove_free_vars_equal_to_args = Ocaml_shadow.Remove_free_vars_equal_to_args

module Remove_unused_arguments = Ocaml_shadow.Remove_unused_arguments

module Remove_unused_closure_vars = Ocaml_shadow.Remove_unused_closure_vars

module Remove_unused_program_constructs = Ocaml_shadow.Remove_unused_program_constructs

module Runtimedef = Ocaml_shadow.Runtimedef

module Schedgen = Ocaml_shadow.Schedgen

module Scheduling = Ocaml_shadow.Scheduling

module Selectgen = Ocaml_shadow.Selectgen

module Selection = Ocaml_shadow.Selection

module Semantics_of_primitives = Ocaml_shadow.Semantics_of_primitives

module Set_of_closures_id = Ocaml_shadow.Set_of_closures_id

module Set_of_closures_origin = Ocaml_shadow.Set_of_closures_origin

module Share_constants = Ocaml_shadow.Share_constants

module Simple_value_approx = Ocaml_shadow.Simple_value_approx

module Simplif = Ocaml_shadow.Simplif

module Simplify_boxed_integer_ops = Ocaml_shadow.Simplify_boxed_integer_ops

module Simplify_boxed_integer_ops_intf = Ocaml_shadow.Simplify_boxed_integer_ops_intf

module Simplify_common = Ocaml_shadow.Simplify_common

module Simplify_primitives = Ocaml_shadow.Simplify_primitives

module Spill = Ocaml_shadow.Spill

module Split = Ocaml_shadow.Split

module Static_exception = Ocaml_shadow.Static_exception

module Strmatch = Ocaml_shadow.Strmatch

module Strongly_connected_components = Ocaml_shadow.Strongly_connected_components

module Stypes = Ocaml_shadow.Stypes

module Subst = Ocaml_shadow.Subst

module Switch = Ocaml_shadow.Switch

module Symbol = Ocaml_shadow.Symbol

module Symtable = Ocaml_shadow.Symtable

module Tag = Ocaml_shadow.Tag

module Targetint = Ocaml_shadow.Targetint

module Tast_iterator = Ocaml_shadow.Tast_iterator

module Tast_mapper = Ocaml_shadow.Tast_mapper

module Terminfo = Ocaml_shadow.Terminfo

module Topdirs = Ocaml_shadow.Topdirs

module Toploop = Ocaml_shadow.Toploop

module Topmain = Ocaml_shadow.Topmain

module Topstart = Ocaml_shadow.Topstart

module Trace = Ocaml_shadow.Trace

module Translattribute = Ocaml_shadow.Translattribute

module Translclass = Ocaml_shadow.Translclass

module Translcore = Ocaml_shadow.Translcore

module Translmod = Ocaml_shadow.Translmod

module Translobj = Ocaml_shadow.Translobj

module Translprim = Ocaml_shadow.Translprim

module Traverse_for_exported_symbols = Ocaml_shadow.Traverse_for_exported_symbols

module Type_immediacy = Ocaml_shadow.Type_immediacy

module Typeclass = Ocaml_shadow.Typeclass

module Typecore = Ocaml_shadow.Typecore

module Typedecl = Ocaml_shadow.Typedecl

module Typedecl_immediacy = Ocaml_shadow.Typedecl_immediacy

module Typedecl_properties = Ocaml_shadow.Typedecl_properties

module Typedecl_separability = Ocaml_shadow.Typedecl_separability

module Typedecl_unboxed = Ocaml_shadow.Typedecl_unboxed

module Typedecl_variance = Ocaml_shadow.Typedecl_variance

module Typedtree = Ocaml_shadow.Typedtree

module Typemod = Ocaml_shadow.Typemod

module Typeopt = Ocaml_shadow.Typeopt

module Types = Ocaml_shadow.Types

module Typetexp = Ocaml_shadow.Typetexp

module Un_anf = Ocaml_shadow.Un_anf

module Unbox_closures = Ocaml_shadow.Unbox_closures

module Unbox_free_vars_of_closures = Ocaml_shadow.Unbox_free_vars_of_closures

module Unbox_specialised_args = Ocaml_shadow.Unbox_specialised_args

module Untypeast = Ocaml_shadow.Untypeast

module Var_within_closure = Ocaml_shadow.Var_within_closure

module Variable = Ocaml_shadow.Variable

module Warnings = Ocaml_shadow.Warnings

module X86_ast = Ocaml_shadow.X86_ast

module X86_dsl = Ocaml_shadow.X86_dsl

module X86_gas = Ocaml_shadow.X86_gas

module X86_masm = Ocaml_shadow.X86_masm

module X86_proc = Ocaml_shadow.X86_proc

Includes the overrides from Ppxlib_ast, as well as all the Ast definitions since we need them in every single ppx

include module type of struct include Ppxlib_ast end

module type OCaml_version = Ppxlib_ast.OCaml_version

module Ast = Ppxlib_ast.Ast

module Ast_helper = Ppxlib_ast.Ast_helper

module Ast_magic = Ppxlib_ast.Ast_magic

module Asttypes = Ppxlib_ast.Asttypes

module Compiler_version = Ppxlib_ast.Compiler_version

module Js = Ppxlib_ast.Js

module Find_version = Ppxlib_ast.Find_version

module Convert = Ppxlib_ast.Convert

module Extra_warnings = Ppxlib_ast.Extra_warnings

module Lexer = Ppxlib_ast.Lexer

module Location_error = Ppxlib_ast.Location_error

module Parse = Ppxlib_ast.Parse

module Parser = Ppxlib_ast.Parser

module Parsetree = Ppxlib_ast.Parsetree

module Pprintast = Ppxlib_ast.Pprintast

module Select_ast = Ppxlib_ast.Select_ast

module Selected_ast = Ppxlib_ast.Selected_ast

module Syntaxerr = Ppxlib_ast.Syntaxerr

module Import_for_core = Ppxlib_ast.Import_for_core

include module type of struct include Ast end

type position = Lexing.position = {

pos_fname : string;

pos_lnum : int;

pos_bol : int;

pos_cnum : int;

}

and location = Warnings.loc = {

loc_start : position;

loc_end : position;

loc_ghost : bool;

}

and location_stack = location list

and 'a loc = 'a Location.loc = {

txt : 'a;

loc : location;

}

and longident = Ocaml_common.Longident.t =

| Lident of string

| Ldot of longident * string

| Lapply of longident * longident

and longident_loc = longident loc

Auxiliary AST types used by parsetree and typedtree.

and rec_flag = Migrate_parsetree.Ast_412.Asttypes.rec_flag =

| Nonrecursive

| Recursive

and direction_flag = Migrate_parsetree.Ast_412.Asttypes.direction_flag =

| Upto

| Downto

and private_flag = Migrate_parsetree.Ast_412.Asttypes.private_flag =

| Private

| Public

and mutable_flag = Migrate_parsetree.Ast_412.Asttypes.mutable_flag =

| Immutable

| Mutable

and virtual_flag = Migrate_parsetree.Ast_412.Asttypes.virtual_flag =

| Virtual

| Concrete

and override_flag = Migrate_parsetree.Ast_412.Asttypes.override_flag =

| Override

| Fresh

and closed_flag = Migrate_parsetree.Ast_412.Asttypes.closed_flag =

| Closed

| Open

and label = string

and arg_label = Migrate_parsetree.Ast_412.Asttypes.arg_label =

| Nolabel

| Labelled of string

| Optional of string

and variance = Migrate_parsetree.Ast_412.Asttypes.variance =

| Covariant

| Contravariant

| NoVariance

and injectivity = Migrate_parsetree.Ast_412.Asttypes.injectivity =

| Injective

| NoInjectivity

Abstract syntax tree produced by parsing

and constant = Migrate_parsetree.Ast_412.Parsetree.constant =

| Pconst_integer of string * char option

| Pconst_char of char

| Pconst_string of string * location * string option

| Pconst_float of string * char option

Extension points

and attribute = Migrate_parsetree.Ast_412.Parsetree.attribute = {

attr_name : string loc;

attr_payload : payload;

attr_loc : location;

}

and extension = string loc * payload

and attributes = attribute list

and payload = Migrate_parsetree.Ast_412.Parsetree.payload =

| PStr of structure

| PSig of signature

| PTyp of core_type

| PPat of pattern * expression option

Core language

and core_type = Migrate_parsetree.Ast_412.Parsetree.core_type = {

ptyp_desc : core_type_desc;

ptyp_loc : location;

ptyp_loc_stack : location_stack;

ptyp_attributes : attributes;

}

and core_type_desc = Migrate_parsetree.Ast_412.Parsetree.core_type_desc =

| Ptyp_any

| Ptyp_var of string

| Ptyp_arrow of arg_label * core_type * core_type

| Ptyp_tuple of core_type list

| Ptyp_constr of longident_loc * core_type list

| Ptyp_object of object_field list * closed_flag

| Ptyp_class of longident_loc * core_type list

| Ptyp_alias of core_type * string

| Ptyp_variant of row_field list * closed_flag * label list option

| Ptyp_poly of string loc list * core_type

| Ptyp_package of package_type

| Ptyp_extension of extension

and package_type = longident_loc * (longident_loc * core_type) list

and row_field = Migrate_parsetree.Ast_412.Parsetree.row_field = {

prf_desc : row_field_desc;

prf_loc : location;

prf_attributes : attributes;

}

and row_field_desc = Migrate_parsetree.Ast_412.Parsetree.row_field_desc =

| Rtag of label loc * bool * core_type list

| Rinherit of core_type

and object_field = Migrate_parsetree.Ast_412.Parsetree.object_field = {

pof_desc : object_field_desc;

pof_loc : location;

pof_attributes : attributes;

}

and object_field_desc = Migrate_parsetree.Ast_412.Parsetree.object_field_desc =

| Otag of label loc * core_type

| Oinherit of core_type

and pattern = Migrate_parsetree.Ast_412.Parsetree.pattern = {

ppat_desc : pattern_desc;

ppat_loc : location;

ppat_loc_stack : location_stack;

ppat_attributes : attributes;

}

and pattern_desc = Migrate_parsetree.Ast_412.Parsetree.pattern_desc =

| Ppat_any

| Ppat_var of string loc

| Ppat_alias of pattern * string loc

| Ppat_constant of constant

| Ppat_interval of constant * constant

| Ppat_tuple of pattern list

| Ppat_construct of longident_loc * pattern option

| Ppat_variant of label * pattern option

| Ppat_record of (longident_loc * pattern) list * closed_flag

| Ppat_array of pattern list

| Ppat_or of pattern * pattern

| Ppat_constraint of pattern * core_type

| Ppat_type of longident_loc

| Ppat_lazy of pattern

| Ppat_unpack of string option loc

| Ppat_exception of pattern

| Ppat_extension of extension

| Ppat_open of longident_loc * pattern

and expression = Migrate_parsetree.Ast_412.Parsetree.expression = {

pexp_desc : expression_desc;

pexp_loc : location;

pexp_loc_stack : location_stack;

pexp_attributes : attributes;

}

and expression_desc = Migrate_parsetree.Ast_412.Parsetree.expression_desc =

| Pexp_ident of longident_loc

| Pexp_constant of constant

| Pexp_let of rec_flag * value_binding list * expression

| Pexp_function of cases

| Pexp_fun of arg_label * expression option * pattern * expression

| Pexp_apply of expression * (arg_label * expression) list

| Pexp_match of expression * cases

| Pexp_try of expression * cases

| Pexp_tuple of expression list

| Pexp_construct of longident_loc * expression option

| Pexp_variant of label * expression option

| Pexp_record of (longident_loc * expression) list * expression option

| Pexp_field of expression * longident_loc

| Pexp_setfield of expression * longident_loc * expression

| Pexp_array of expression list

| Pexp_ifthenelse of expression * expression * expression option

| Pexp_sequence of expression * expression

| Pexp_while of expression * expression

| Pexp_for of pattern * expression * expression * direction_flag * expression

| Pexp_constraint of expression * core_type

| Pexp_coerce of expression * core_type option * core_type

| Pexp_send of expression * label loc

| Pexp_new of longident_loc

| Pexp_setinstvar of label loc * expression

| Pexp_override of (label loc * expression) list

| Pexp_letmodule of string option loc * module_expr * expression

| Pexp_letexception of extension_constructor * expression

| Pexp_assert of expression

| Pexp_lazy of expression

| Pexp_poly of expression * core_type option

| Pexp_object of class_structure

| Pexp_newtype of string loc * expression

| Pexp_pack of module_expr

| Pexp_open of open_declaration * expression

| Pexp_letop of letop

| Pexp_extension of extension

| Pexp_unreachable

and case = Migrate_parsetree.Ast_412.Parsetree.case = {

pc_lhs : pattern;

pc_guard : expression option;

pc_rhs : expression;

}

and letop = Migrate_parsetree.Ast_412.Parsetree.letop = {

let_ : binding_op;

ands : binding_op list;

body : expression;

}

and binding_op = Migrate_parsetree.Ast_412.Parsetree.binding_op = {

pbop_op : string loc;

pbop_pat : pattern;

pbop_exp : expression;

pbop_loc : location;

}

and value_description = Migrate_parsetree.Ast_412.Parsetree.value_description = {

pval_name : string loc;

pval_type : core_type;

pval_prim : string list;

pval_attributes : attributes;

pval_loc : location;

}

and type_declaration = Migrate_parsetree.Ast_412.Parsetree.type_declaration = {

ptype_name : string loc;

ptype_params : (core_type * (variance * injectivity)) list;

ptype_cstrs : (core_type * core_type * location) list;

ptype_kind : type_kind;

ptype_private : private_flag;

ptype_manifest : core_type option;

ptype_attributes : attributes;

ptype_loc : location;

}

and type_kind = Migrate_parsetree.Ast_412.Parsetree.type_kind =

| Ptype_abstract

| Ptype_variant of constructor_declaration list

| Ptype_record of label_declaration list

| Ptype_open

and label_declaration = Migrate_parsetree.Ast_412.Parsetree.label_declaration = {

pld_name : string loc;

pld_mutable : mutable_flag;

pld_type : core_type;

pld_loc : location;

pld_attributes : attributes;

}

and constructor_declaration = Migrate_parsetree.Ast_412.Parsetree.constructor_declaration = {

pcd_name : string loc;

pcd_args : constructor_arguments;

pcd_res : core_type option;

pcd_loc : location;

pcd_attributes : attributes;

}

and constructor_arguments = Migrate_parsetree.Ast_412.Parsetree.constructor_arguments =

| Pcstr_tuple of core_type list

| Pcstr_record of label_declaration list

and type_extension = Migrate_parsetree.Ast_412.Parsetree.type_extension = {

ptyext_path : longident_loc;

ptyext_params : (core_type * (variance * injectivity)) list;

ptyext_constructors : extension_constructor list;

ptyext_private : private_flag;

ptyext_loc : location;

ptyext_attributes : attributes;

}

and extension_constructor = Migrate_parsetree.Ast_412.Parsetree.extension_constructor = {

pext_name : string loc;

pext_kind : extension_constructor_kind;

pext_loc : location;

pext_attributes : attributes;

}

and type_exception = Migrate_parsetree.Ast_412.Parsetree.type_exception = {

ptyexn_constructor : extension_constructor;

ptyexn_loc : location;

ptyexn_attributes : attributes;

}

and extension_constructor_kind = Migrate_parsetree.Ast_412.Parsetree.extension_constructor_kind =

| Pext_decl of constructor_arguments * core_type option

| Pext_rebind of longident_loc

Class language

and class_type = Migrate_parsetree.Ast_412.Parsetree.class_type = {

pcty_desc : class_type_desc;

pcty_loc : location;

pcty_attributes : attributes;

}

and class_type_desc = Migrate_parsetree.Ast_412.Parsetree.class_type_desc =

| Pcty_constr of longident_loc * core_type list

| Pcty_signature of class_signature

| Pcty_arrow of arg_label * core_type * class_type

| Pcty_extension of extension

| Pcty_open of open_description * class_type

and class_signature = Migrate_parsetree.Ast_412.Parsetree.class_signature = {

pcsig_self : core_type;

pcsig_fields : class_type_field list;

}

and class_type_field = Migrate_parsetree.Ast_412.Parsetree.class_type_field = {

pctf_desc : class_type_field_desc;

pctf_loc : location;

pctf_attributes : attributes;

}

and class_type_field_desc = Migrate_parsetree.Ast_412.Parsetree.class_type_field_desc =

| Pctf_inherit of class_type

| Pctf_val of label loc * mutable_flag * virtual_flag * core_type

| Pctf_method of label loc * private_flag * virtual_flag * core_type

| Pctf_constraint of core_type * core_type

| Pctf_attribute of attribute

| Pctf_extension of extension

and 'a class_infos = 'a Migrate_parsetree.Ast_412.Parsetree.class_infos = {

pci_virt : virtual_flag;

pci_params : (core_type * (variance * injectivity)) list;

pci_name : string loc;

pci_expr : 'a;

pci_loc : location;

pci_attributes : attributes;

}

and class_description = class_type class_infos

and class_type_declaration = class_type class_infos

and class_expr = Migrate_parsetree.Ast_412.Parsetree.class_expr = {

pcl_desc : class_expr_desc;

pcl_loc : location;

pcl_attributes : attributes;

}

and class_expr_desc = Migrate_parsetree.Ast_412.Parsetree.class_expr_desc =

| Pcl_constr of longident_loc * core_type list

| Pcl_structure of class_structure

| Pcl_fun of arg_label * expression option * pattern * class_expr

| Pcl_apply of class_expr * (arg_label * expression) list

| Pcl_let of rec_flag * value_binding list * class_expr

| Pcl_constraint of class_expr * class_type

| Pcl_extension of extension

| Pcl_open of open_description * class_expr

and class_structure = Migrate_parsetree.Ast_412.Parsetree.class_structure = {

pcstr_self : pattern;

pcstr_fields : class_field list;

}

and class_field = Migrate_parsetree.Ast_412.Parsetree.class_field = {

pcf_desc : class_field_desc;

pcf_loc : location;

pcf_attributes : attributes;

}

and class_field_desc = Migrate_parsetree.Ast_412.Parsetree.class_field_desc =

| Pcf_inherit of override_flag * class_expr * string loc option

| Pcf_val of label loc * mutable_flag * class_field_kind

| Pcf_method of label loc * private_flag * class_field_kind

| Pcf_constraint of core_type * core_type

| Pcf_initializer of expression

| Pcf_attribute of attribute

| Pcf_extension of extension

and class_field_kind = Migrate_parsetree.Ast_412.Parsetree.class_field_kind =

| Cfk_virtual of core_type

| Cfk_concrete of override_flag * expression

and class_declaration = class_expr class_infos

Module language

and module_type = Migrate_parsetree.Ast_412.Parsetree.module_type = {

pmty_desc : module_type_desc;

pmty_loc : location;

pmty_attributes : attributes;

}

and module_type_desc = Migrate_parsetree.Ast_412.Parsetree.module_type_desc =

| Pmty_ident of longident_loc

| Pmty_signature of signature

| Pmty_functor of functor_parameter * module_type

| Pmty_with of module_type * with_constraint list

| Pmty_typeof of module_expr

| Pmty_extension of extension

| Pmty_alias of longident_loc

and functor_parameter = Migrate_parsetree.Ast_412.Parsetree.functor_parameter =

| Unit

| Named of string option loc * module_type

and signature = signature_item list

and signature_item = Migrate_parsetree.Ast_412.Parsetree.signature_item = {

psig_desc : signature_item_desc;

psig_loc : location;

}

and signature_item_desc = Migrate_parsetree.Ast_412.Parsetree.signature_item_desc =

| Psig_value of value_description

| Psig_type of rec_flag * type_declaration list

| Psig_typesubst of type_declaration list

| Psig_typext of type_extension

| Psig_exception of type_exception

| Psig_module of module_declaration

| Psig_modsubst of module_substitution

| Psig_recmodule of module_declaration list

| Psig_modtype of module_type_declaration

| Psig_open of open_description

| Psig_include of include_description

| Psig_class of class_description list

| Psig_class_type of class_type_declaration list

| Psig_attribute of attribute

| Psig_extension of extension * attributes

and module_declaration = Migrate_parsetree.Ast_412.Parsetree.module_declaration = {

pmd_name : string option loc;

pmd_type : module_type;

pmd_attributes : attributes;

pmd_loc : location;

}

and module_substitution = Migrate_parsetree.Ast_412.Parsetree.module_substitution = {

pms_name : string loc;

pms_manifest : longident_loc;

pms_attributes : attributes;

pms_loc : location;

}

and module_type_declaration = Migrate_parsetree.Ast_412.Parsetree.module_type_declaration = {

pmtd_name : string loc;

pmtd_type : module_type option;

pmtd_attributes : attributes;

pmtd_loc : location;

}

and 'a open_infos = 'a Migrate_parsetree.Ast_412.Parsetree.open_infos = {

popen_expr : 'a;

popen_override : override_flag;

popen_loc : location;

popen_attributes : attributes;

}

and open_description = longident_loc open_infos

and open_declaration = module_expr open_infos

and 'a include_infos = 'a Migrate_parsetree.Ast_412.Parsetree.include_infos = {

pincl_mod : 'a;

pincl_loc : location;

pincl_attributes : attributes;

}

and include_description = module_type include_infos

and include_declaration = module_expr include_infos

and with_constraint = Migrate_parsetree.Ast_412.Parsetree.with_constraint =

| Pwith_type of longident_loc * type_declaration

| Pwith_module of longident_loc * longident_loc

| Pwith_typesubst of longident_loc * type_declaration

| Pwith_modsubst of longident_loc * longident_loc

and module_expr = Migrate_parsetree.Ast_412.Parsetree.module_expr = {

pmod_desc : module_expr_desc;

pmod_loc : location;

pmod_attributes : attributes;

}

and module_expr_desc = Migrate_parsetree.Ast_412.Parsetree.module_expr_desc =

| Pmod_ident of longident_loc

| Pmod_structure of structure

| Pmod_functor of functor_parameter * module_expr

| Pmod_apply of module_expr * module_expr

| Pmod_constraint of module_expr * module_type

| Pmod_unpack of expression

| Pmod_extension of extension

and structure = structure_item list

and structure_item = Migrate_parsetree.Ast_412.Parsetree.structure_item = {

pstr_desc : structure_item_desc;

pstr_loc : location;

}

and structure_item_desc = Migrate_parsetree.Ast_412.Parsetree.structure_item_desc =

| Pstr_eval of expression * attributes

| Pstr_value of rec_flag * value_binding list

| Pstr_primitive of value_description

| Pstr_type of rec_flag * type_declaration list

| Pstr_typext of type_extension

| Pstr_exception of type_exception

| Pstr_module of module_binding

| Pstr_recmodule of module_binding list

| Pstr_modtype of module_type_declaration

| Pstr_open of open_declaration

| Pstr_class of class_declaration list

| Pstr_class_type of class_type_declaration list

| Pstr_include of include_declaration

| Pstr_attribute of attribute

| Pstr_extension of extension * attributes

and value_binding = Migrate_parsetree.Ast_412.Parsetree.value_binding = {

pvb_pat : pattern;

pvb_expr : expression;

pvb_attributes : attributes;

pvb_loc : location;

}

and module_binding = Migrate_parsetree.Ast_412.Parsetree.module_binding = {

pmb_name : string option loc;

pmb_expr : module_expr;

pmb_attributes : attributes;

pmb_loc : location;

}

Toplevel

and toplevel_phrase = Migrate_parsetree.Ast_412.Parsetree.toplevel_phrase =

| Ptop_def of structure

| Ptop_dir of toplevel_directive

and toplevel_directive = Migrate_parsetree.Ast_412.Parsetree.toplevel_directive = {

pdir_name : string loc;

pdir_arg : directive_argument option;

pdir_loc : location;

}

and directive_argument = Migrate_parsetree.Ast_412.Parsetree.directive_argument = {

pdira_desc : directive_argument_desc;

pdira_loc : location;

}

and directive_argument_desc = Migrate_parsetree.Ast_412.Parsetree.directive_argument_desc =

| Pdir_string of string

| Pdir_int of string * char option

| Pdir_ident of longident

| Pdir_bool of bool

and cases = case list

class virtual map : object ... end

class virtual iter : object ... end

class virtual +'b fold : object ... end

class virtual +'b fold_map : object ... end

class virtual +'b map_with_context : object ... end

class virtual +'b lift : object ... end

module Ast_builder : sig ... end

Helpers for build OCaml AST fragments

module Ast_pattern : sig ... end

First class AST patterns

module Ast_traverse : sig ... end

AST traversal classes

module Attribute : sig ... end

Attribute hygiene

module Code_path : sig ... end

module Caller_id : sig ... end

module Context_free : sig ... end

Context free rewriting

module Deriving : sig ... end

Deriving code from type declarations

module Driver : sig ... end

module Expansion_context : sig ... end

module Extension : sig ... end

module File_path : sig ... end

Return the path used as root in a file

module Loc : sig ... end

Located items

module Location : sig ... end

Overrides the Location module of OCaml

module Longident : sig ... end

Overrides the Longident module of OCaml

module Merlin_helpers : sig ... end

Some helpers to annotate the AST so merlin can decide which branches to look at and which branches to ignore.

module Reserved_namespaces : sig ... end

module Spellcheck : sig ... end

module Quoter : sig ... end

Generate expressions in a hygienic way.

val lident : string -> Longident.t

val core_type_of_type_declaration : Migrate_parsetree.Ast_412.Parsetree.type_declaration -> Migrate_parsetree.Ast_412.Parsetree.core_type

val name_type_params_in_td : Migrate_parsetree.Ast_412.Parsetree.type_declaration -> Migrate_parsetree.Ast_412.Parsetree.type_declaration

val combinator_type_of_type_declaration : Migrate_parsetree.Ast_412.Parsetree.type_declaration -> f:(loc:Location.t -> Migrate_parsetree.Ast_412.Parsetree.core_type -> Migrate_parsetree.Ast_412.Parsetree.core_type) -> Migrate_parsetree.Ast_412.Parsetree.core_type

val gen_symbol : ?prefix:string -> unit -> string

gen_symbol ?prefix () generates a fresh variable name with prefix.

parameter prefix
default = "_x"

val string_of_core_type : Migrate_parsetree.Ast_412.Parsetree.core_type -> string

val assert_no_attributes : Migrate_parsetree.Ast_412.Parsetree.attribute list -> unit

val assert_no_attributes_in : Ast_traverse.iter

val get_type_param_name : (Migrate_parsetree.Ast_412.Parsetree.core_type * (Migrate_parsetree.Ast_412.Asttypes.variance * Migrate_parsetree.Ast_412.Asttypes.injectivity)) -> string Loc.t

get_tparam_id tp returns the string identifier associated with tp if it is a type parameter.

class type_is_recursive : Migrate_parsetree.Ast_412.Asttypes.rec_flag -> Migrate_parsetree.Ast_412.Parsetree.type_declaration list -> object ... end

(new type_is_recursive rec_flag tds)#go () returns whether rec_flag, tds is really a recursive type. We disregard recursive occurrences appearing in arrow types. You can override the search for certain type expressions by inheriting from this class.

val really_recursive : Migrate_parsetree.Ast_412.Asttypes.rec_flag -> Migrate_parsetree.Ast_412.Parsetree.type_declaration list -> Migrate_parsetree.Ast_412.Asttypes.rec_flag

really_recursive rec_flag tds = (new type_is_recursive rec_flag tds)#go ()

val loc_of_payload : Migrate_parsetree.Ast_412.Parsetree.attribute -> Location.t

val loc_of_attribute : Migrate_parsetree.Ast_412.Parsetree.attribute -> Location.t

val loc_of_extension : (string Location.loc * Migrate_parsetree.Ast_412.Parsetree.payload) -> Location.t

val curry_applications : Migrate_parsetree.Ast_412.Parsetree.expression -> Migrate_parsetree.Ast_412.Parsetree.expression

convert multi-arg function applications into a cascade of 1-arg applications

val attribute_of_warning : Location.t -> string -> Migrate_parsetree.Ast_412.Parsetree.attribute

Encode a warning message into an 'ocaml.ppwarning' attribute which can be inserted in a generated Parsetree. The compiler will be responsible for reporting the warning.

val is_polymorphic_variant : Migrate_parsetree.Ast_412.Parsetree.type_declaration -> sig_:bool -> [> `Definitely | `Maybe | `Surely_not ]

val mk_named_sig : loc:Location.t -> sg_name:string -> handle_polymorphic_variant:bool -> Migrate_parsetree.Ast_412.Parsetree.type_declaration list -> Migrate_parsetree.Ast_412.Parsetree.module_type Migrate_parsetree.Ast_412.Parsetree.include_infos option

mk_named_sig ~loc ~sg_name:"Foo" ~handle_polymorphic_variant tds will generate

include Foo (* or Foo1, Foo2, Foo3 *)
  with type (* ('a, 'b, 'c) *) t := (* ('a, 'b, 'c) *) t

when:

there is only one type declaration

the type is named t

there are less than 4 type parameters

there are no constraints on the type parameters

It will take care of giving fresh names to unnamed type parameters.