Module Ppxlib.Deriving

Deriving code from type declarations.

module Args : sig ... end

Specification of generator arguments

Generator registration

type t

Type of registered derivers

module Generator : sig ... end

val add : ?str_type_decl: (Astlib.Ast_500.Parsetree.structure_item list, Astlib.Ast_500.Asttypes.rec_flag * Astlib.Ast_500.Parsetree.type_declaration list) Generator.t -> ?str_type_ext: (Astlib.Ast_500.Parsetree.structure_item list, Astlib.Ast_500.Parsetree.type_extension) Generator.t -> ?str_exception: (Astlib.Ast_500.Parsetree.structure_item list, Astlib.Ast_500.Parsetree.type_exception) Generator.t -> ?str_module_type_decl: (Astlib.Ast_500.Parsetree.structure_item list, Astlib.Ast_500.Parsetree.module_type_declaration) Generator.t -> ?sig_type_decl: (Astlib.Ast_500.Parsetree.signature_item list, Astlib.Ast_500.Asttypes.rec_flag * Astlib.Ast_500.Parsetree.type_declaration list) Generator.t -> ?sig_type_ext: (Astlib.Ast_500.Parsetree.signature_item list, Astlib.Ast_500.Parsetree.type_extension) Generator.t -> ?sig_exception: (Astlib.Ast_500.Parsetree.signature_item list, Astlib.Ast_500.Parsetree.type_exception) Generator.t -> ?sig_module_type_decl: (Astlib.Ast_500.Parsetree.signature_item list, Astlib.Ast_500.Parsetree.module_type_declaration) Generator.t -> ?extension: (loc:Location.t -> path:string -> Astlib.Ast_500.Parsetree.core_type -> Astlib.Ast_500.Parsetree.expression) -> string -> t

Register a new deriving generator.

The various arguments are for the various items on which derivers can be attached in structure and signatures.

We distinguish exception from type_extension as exception E is not exactly the same as type exn += E. Indeed if the type exn is redefined, then type exn += E will add E to the new exn type while exception E will add E to the predefined exn type.

extension register an expander for extension with the name of the deriver. This is here mostly to support the ppx_deriving backend.

val add_alias : string -> ?str_type_decl:t list -> ?str_type_ext:t list -> ?str_exception:t list -> ?str_module_type_decl:t list -> ?sig_type_decl:t list -> ?sig_type_ext:t list -> ?sig_exception:t list -> ?sig_module_type_decl:t list -> t list -> t

add_alias name set add an alias. When the user write the alias, all the generator of set will be used instead. It is possible to override the set for any of the context by passing the specific set in the approriate optional argument of add_alias.

val ignore : t -> unit

Ignore a deriver. So that one can write: Deriving.add ... |> Deriving.ignore