Module Driver.Instrument

type t

type pos =

| Before

| After

val make : (Ppxlib_ast.Parsetree.structure -> Ppxlib_ast.Parsetree.structure) -> position:pos -> t

make transformation ~position creates an instrumentation that can be passed to Driver.register_transformation to instrument an implementation file. transformation is the transformation that will be applied to the AST; position specifies if it should be applied before or after rewriters defined through rules, impl or intf are applied.

module V2 : sig ... end