Module Driver.V2

val register_transformation : ?⁠extensions:Extension.t list -> ?⁠rules:Context_free.Rule.t list -> ?⁠enclose_impl:(Expansion_context.Base.t -> Location.t option -> Ppxlib__.Import.structure * Ppxlib__.Import.structure) -> ?⁠enclose_intf:(Expansion_context.Base.t -> Location.t option -> Ppxlib__.Import.signature * Ppxlib__.Import.signature) -> ?⁠impl:(Expansion_context.Base.t -> Ppxlib__.Import.structure -> Ppxlib__.Import.structure) -> ?⁠intf:(Expansion_context.Base.t -> Ppxlib__.Import.signature -> Ppxlib__.Import.signature) -> ?⁠lint_impl:(Expansion_context.Base.t -> Ppxlib__.Import.structure -> Lint_error.t list) -> ?⁠lint_intf:(Expansion_context.Base.t -> Ppxlib__.Import.signature -> Lint_error.t list) -> ?⁠preprocess_impl:(Expansion_context.Base.t -> Ppxlib__.Import.structure -> Ppxlib__.Import.structure) -> ?⁠preprocess_intf:(Expansion_context.Base.t -> Ppxlib__.Import.signature -> Ppxlib__.Import.signature) -> ?⁠instrument:Instrument.t -> ?⁠aliases:string list -> string -> unit

Same as Driver.register_transformation, but the callbacks have access to an expansion context. Their signatures coincide with the signatures of the respective methods in Ast_traverse.map_with_expansion_context.

val register_transformation_using_ocaml_current_ast : ?⁠impl:(Expansion_context.Base.t -> Ppxlib__.Import.Compiler_version.Ast.Parsetree.structure -> Ppxlib__.Import.Compiler_version.Ast.Parsetree.structure) -> ?⁠intf:(Expansion_context.Base.t -> Ppxlib__.Import.Compiler_version.Ast.Parsetree.signature -> Ppxlib__.Import.Compiler_version.Ast.Parsetree.signature) -> ?⁠aliases:string list -> string -> unit

Same as Driver.register_transformation_using_ocaml_current_ast, but the callbacks ?impl and ?intf have access to an expansion context.