Module Expansion_context.Extension

type t

Type of expansion contexts for extensions

val extension_point_loc : t -> Location.t

Return the location of the extension point being expanded

val code_path : t -> Code_path.t

Return the code path for the given context

val tool_name : t -> string

Can be used within a ppx preprocessor to know which tool is calling it "ocamlc", "ocamlopt", "ocamldep", "ocaml", ... .

val with_loc_and_path : (loc:Location.t -> path:string -> 'a) -> ctxt:t -> 'a

Wrap a fun ~loc ~path into a fun ~ctxt