Module Ppxlib.Location

Overrides the Location module of OCaml

There are less functions in this module. However the API should be more stable than the Location module of OCaml.

type t = Warnings.loc = {

loc_start : Lexing.position;

loc_end : Lexing.position;

loc_ghost : bool;

}

val in_file : string -> t

Return an empty ghost range located in a given file.

val none : t

An arbitrary value of type t; describes an empty ghost range.

val init : Lexing.lexbuf -> string -> unit

Set the file name and line number of the lexbuf to be the start of the named file.

val raise_errorf : ?loc:t -> ('a, Stdppx.Caml.Format.formatter, unit, 'b) format4 -> 'a

Raise a located error. The exception is caught by driver and handled appropriately

val of_lexbuf : Lexing.lexbuf -> t

Return the location corresponding to the last matched regular expression

val report_exception : Stdppx.Caml.Format.formatter -> exn -> unit

Report an exception on the given formatter

val print : Stdppx.Caml.Format.formatter -> t -> unit

Prints File "...", line ..., characters ...-...:

type nonrec 'a loc = 'a Location.loc = {

txt : 'a;

loc : t;

}

val compare_pos : Lexing.position -> Lexing.position -> int

val min_pos : Lexing.position -> Lexing.position -> Lexing.position

val max_pos : Lexing.position -> Lexing.position -> Lexing.position

val compare : t -> t -> int

module Error : sig ... end

exception Error of Error.t