Module Location.Error

type t

val make : loc:t -> string -> sub:(t * string) list -> t

val createf : loc:t -> ('a, unit, string, t) format4 -> 'a

val message : t -> string

val set_message : t -> string -> t

val register_error_of_exn : (exn -> t option) -> unit

Register an exception handler. Exception registered this way will be properly displayed by report_exception.

val of_exn : exn -> t option

val to_extension : t -> string Location.loc * Migrate_parsetree.Ast_412.Parsetree.payload

Convert an error to an extension point. The compiler recognizes this and displays the error properly.

val raise : t -> 'a

Raise a compiler Parsing.Location.Error exception. The composition of Location.Error.createf with Location.Error.raise is the same as Location.raise_errorf.

val update_loc : t -> t -> t